requirements.txt
flask
requests
psutil
python-dotenv
redis